requirements.txt
pymysql
python-dotenv
